Rename files; update references
