+++ sdk/storage/azure-storage-blob/tests/settings/testcase.py
@@ -9,8 +9,8 @@ from __future__ import division
 import functools
 import os
 import logging
-from devtools_testutils import (
-    PowerShellPreparer, ACCOUNT_FAKE_KEY)
+from devtools_testutils import PowerShellPreparer
+from devtools_testutils.fake_credentials import STORAGE_ACCOUNT_FAKE_KEY
 
 try:
     from cStringIO import StringIO      # Python 2
@@ -51,15 +51,15 @@ os.environ['STORAGE_CLIENT_SECRET'] = os.environ.get('STORAGE_CLIENT_SECRET', No
 BlobPreparer = functools.partial(
     PowerShellPreparer, "storage",
     storage_account_name="storagename",
-    storage_account_key=ACCOUNT_FAKE_KEY,
+    storage_account_key=STORAGE_ACCOUNT_FAKE_KEY,
     secondary_storage_account_name="pyrmtstoragestorname",
-    secondary_storage_account_key=ACCOUNT_FAKE_KEY,
+    secondary_storage_account_key=STORAGE_ACCOUNT_FAKE_KEY,
     blob_storage_account_name="storagenamestorname",
-    blob_storage_account_key=ACCOUNT_FAKE_KEY,
+    blob_storage_account_key=STORAGE_ACCOUNT_FAKE_KEY,
     versioned_storage_account_name="storagenamestorname",
-    versioned_storage_account_key=ACCOUNT_FAKE_KEY,
+    versioned_storage_account_key=STORAGE_ACCOUNT_FAKE_KEY,
     premium_storage_account_name='pyacrstoragestorname',
-    premium_storage_account_key=ACCOUNT_FAKE_KEY,
+    premium_storage_account_key=STORAGE_ACCOUNT_FAKE_KEY,
     storage_resource_group_name="rgname",
 
 )
