+++ sdk/storage/azure-storage-file-share/tests/settings/testcase.py
@@ -15,8 +15,8 @@ except ImportError:
     import mock
 
 import logging
-from devtools_testutils import (
-    PowerShellPreparer, ACCOUNT_FAKE_KEY)
+from devtools_testutils import PowerShellPreparer
+from devtools_testutils.fake_credentials import STORAGE_ACCOUNT_FAKE_KEY
 try:
     from cStringIO import StringIO      # Python 2
 except ImportError:
@@ -57,11 +57,11 @@ os.environ['STORAGE_CLIENT_SECRET'] = os.environ.get('STORAGE_CLIENT_SECRET', No
 FileSharePreparer = functools.partial(
     PowerShellPreparer, "storage",
     storage_account_name="storagename",
-    storage_account_key=ACCOUNT_FAKE_KEY,
+    storage_account_key=STORAGE_ACCOUNT_FAKE_KEY,
     premium_storage_file_account_name="pyacrstoragestorname",
-    premium_storage_file_account_key=ACCOUNT_FAKE_KEY,
+    premium_storage_file_account_key=STORAGE_ACCOUNT_FAKE_KEY,
     secondary_storage_account_name="pyrmtstoragestorname",
-    secondary_storage_account_key=ACCOUNT_FAKE_KEY,
+    secondary_storage_account_key=STORAGE_ACCOUNT_FAKE_KEY,
 )
