+++ tools/azure-sdk-tools/devtools_testutils/azure_recorded_testcase.py
@@ -22,7 +22,7 @@ from .sanitizers import add_general_regex_sanitizer
 
 try:
     # Try to import the AsyncFakeCredential, if we cannot assume it is Python 2
-    from .fake_async_credential import AsyncFakeCredential
+    from .fake_credentials_async import AsyncFakeCredential
 except SyntaxError:
     pass
