+++ tools/azure-sdk-tools/devtools_testutils/fake_credentials_async.py
